+++ src/matrix/mod.rs
@@ -3,3 +3,4 @@ pub mod mat;
 pub mod mat_view;
 mod mul;
 mod sub;
+mod transpose;
